ENG-02: split analysis.py _run_* into experiments/_analyses
Move the 13 _run_* analyses and the 2 _compute_* summary helpers into
focused modules under experiments/_analyses/ (ols_extractors, diagnostics,
lack_of_fit, curvature, model_selection, box_cox, lenth, prediction,
_shared). analysis.py keeps build_formula, AnalysisResult and the
analyze_experiment dispatcher, re-exporting the implementations (so
_run_residual_diagnostics stays importable for tests/test_sec09).
Behaviour and analyze_experiment output verified byte-identical.
